Remove feed callback flow

+++ routes/api.php
@@ -12,18 +12,12 @@ Route::post('/register', [AuthController::class, 'register']);
 Route::post('/login', [AuthController::class, 'login']);
 
 Route::middleware('auth:sanctum')->get('/user', [AuthenticatedUserController::class, 'show']);
-
 Route::get('/posts/{post}', [PostController::class, 'show']);
 
-Route::middleware('auth:sanctum')->group(function () {
-    Route::get('/feed', [FeedController::class, 'index']);
-
-    Route::post('/follow/{userToFollow}', [FollowController::class, 'store']);
-    Route::delete('/follow/{userToUnfollow}', [FollowController::class, 'destroy']);
-
-    Route::post('/posts', [PostController::class, 'store']);
-    Route::delete('/posts/{post}', [PostController::class, 'destroy']);
-
-    Route::post('/posts/{post}/like', [LikeController::class, 'store']);
-    Route::delete('/posts/{post}/like', [LikeController::class, 'destroy']);
-});
+Route::middleware('auth:sanctum')->get('/feed', [FeedController::class, 'index']);
+Route::middleware('auth:sanctum')->post('/follow/{userToFollow}', [FollowController::class, 'store']);
+Route::middleware('auth:sanctum')->delete('/follow/{userToUnfollow}', [FollowController::class, 'destroy']);
+Route::middleware('auth:sanctum')->post('/posts', [PostController::class, 'store']);
+Route::middleware('auth:sanctum')->delete('/posts/{post}', [PostController::class, 'destroy']);
+Route::middleware('auth:sanctum')->post('/posts/{post}/like', [LikeController::class, 'store']);
+Route::middleware('auth:sanctum')->delete('/posts/{post}/like', [LikeController::class, 'destroy']);
